+++ src/modules/restaurant/dto/restaurant.dto.ts
@@ -1,0 +1,16 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { IsString } from 'class-validator';
+
+export class SearchRestaurantsDto {
+  @IsString()
+  @ApiProperty()
+  radius?: number;
+
+  @IsString()
+  @ApiProperty()
+  latitude?: number;
+
+  @IsString()
+  @ApiProperty()
+  longitude?: number;
+}