+++ src/modules/restaurant/restaurant.controller.ts
@@ -1,10 +1,10 @@
 import {
-  Body,
   Controller,
   Get,
   UseGuards,
   Version,
   Request,
+  Query,
 } from '@nestjs/common';
 import {
   ApiBadRequestResponse,
@@ -19,10 +19,20 @@ import {
   UNAUTHORIZED_ERROR,
 } from 'src/shared/errors/response-errors';
 import { JwtAuthGuard } from 'src/shared/guards/jwt-auth.guard';
+import { RestaurantService } from './restaurant.service';
+import { SearchRestaurantsDto } from './dto/restaurant.dto';
+import { RestaurantHistory } from 'src/db/restaurant/entities/restaurant-history.entity';
+import {
+  CustomException,
+  CustomInternalServerException,
+} from 'src/shared/errors/custom-exceptions';
+import { ErrorCodesEnum } from 'src/shared/errors/erro-codes.enum';
 
 @Controller('restaurant')
 @ApiTags('restaurant')
 export class RestaurantController {
+  constructor(private readonly restaurantService: RestaurantService) {}
+
   @UseGuards(JwtAuthGuard)
   @Version('1')
   @ApiOperation({
@@ -39,8 +49,26 @@ export class RestaurantController {
     description: UNAUTHORIZED_ERROR,
   })
   @Get()
-  async findByCoordinates(@Body() hello: string): Promise<string> {
-    return `${hello}`;
+  async findByCoordinates(
+    @Query()
+    { radius, latitude, longitude }: SearchRestaurantsDto,
+    @Request() req,
+  ): Promise<any> {
+    try {
+      const { sub } = req.user;
+      const result = await this.restaurantService.searchByCity(
+        sub,
+        radius,
+        latitude,
+        longitude,
+      );
+      return result;
+    } catch (error) {
+      if (error instanceof CustomException) {
+        throw error;
+      }
+      throw new CustomInternalServerException(ErrorCodesEnum.TBA016, error);
+    }
   }
 
   @UseGuards(JwtAuthGuard)
@@ -59,8 +87,15 @@ export class RestaurantController {
     description: UNAUTHORIZED_ERROR,
   })
   @Get('history')
-  async findHistory(@Request() req): Promise<string> {
-    const { username } = req.user;
-    return `Hola ${username}`;
+  async findHistory(@Request() req): Promise<RestaurantHistory[]> {
+    try {
+      const { sub } = req.user;
+      return this.restaurantService.getHistoryByUser(sub);
+    } catch (error) {
+      if (error instanceof CustomException) {
+        throw error;
+      }
+      throw new CustomInternalServerException(ErrorCodesEnum.TBA015, error);
+    }
   }
 }