+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Add connection with postgres and basic configuration and migrations
 - Add create user in db functionality
 - Add sign in and sign up functionality
+- Add search restaurants and save history